Advanced coupled matrix and tensor factorizations

Advanced coupled matrix and tensor factorizations

Usage

acmtf_opt(
  Z,
  numComponents,
  initialization = "random",
  alpha = 1,
  beta = rep(0.001, length(Z$object)),
  epsilon = 1e-08,
  cg_update = "HS",
  line_search = "MT",
  max_iter = 10000,
  max_fn = 10000,
  abs_tol = 1e-10,
  rel_tol = 1e-10,
  grad_tol = 1e-10,
  nstart = 1,
  numCores = 1,
  sortComponents = TRUE,
  allOutput = FALSE
)

Arguments

Z

Combined dataset and mode object as produced by setupCMTFdata().

numComponents

Number of components

initialization

Initialization, either "random" (default) or "nvec" for numComponents components of the concatenated data using svd.

alpha

Scalar penalizing the components to be norm 1 (default 1).

beta

Vector of penalty values for each dataset, penalizing the lambda terms (default 1e-3).

epsilon

Scalar value to make it possible to compute the partial derivatives of lambda (default 1e-8).

cg_update

Update method for the conjugate gradient algorithm, see mize::mize() for the options (default="HS", Hestenes-Steifel).

line_search

Line search algorithm to use, see mize::mize() for the options (default="MT", More-Thuente).

max_iter

Maximum number of iterations.

max_fn

Maximum number of function evaluations.

abs_tol

Function tolerance criterion for convergence.

rel_tol

Relative function tolerance criterion for convergence.

grad_tol

Absolute tolerence for the l2-norm of the gradient vector.

nstart

Number of models to produce (default 1). If set higher than one, the package will return the best fitted model.

numCores

Number of cores to use (default 1). If set higher than one, the package will attempt to run in parallel.

sortComponents

Sort the components in the output by descending order of variation explained.

allOutput

Return all created models. Ignored if nstart=1.

Value

List object, similar to mize::mize() output. Includes a Fac object of the model, which is a list of components per mode. Also includes an init object giving the initialized input vectors.

Examples

A = array(rnorm(108*2), c(108, 2))
B = array(rnorm(100*2), c(100, 2))
C = array(rnorm(10*2), c(10, 2))
D = array(rnorm(100*2), c(100,2))
E = array(rnorm(10*2), c(10,2))

df1 = reinflateTensor(A, B, C)
df2 = reinflateTensor(A, D, E)
datasets = list(df1, df2)
modes = list(c(1,2,3), c(1,4,5))
Z = setupCMTFdata(datasets, modes, normalize=FALSE)

# specific setting to reduce runtime for CRAN
model = acmtf_opt(Z, 1, rel_tol=1e-5, abs_tol=1e-5)
#> [[1]]
#>               [,1]
#>   [1,] -0.63819643
#>   [2,]  0.77470821
#>   [3,] -0.77606625
#>   [4,] -1.99961818
#>   [5,] -3.48287833
#>   [6,] -1.22741232
#>   [7,] -0.93894324
#>   [8,]  0.97689142
#>   [9,]  0.60192778
#>  [10,]  1.95424746
#>  [11,]  2.18199623
#>  [12,] -0.78857854
#>  [13,]  1.88703486
#>  [14,]  7.52249812
#>  [15,]  4.22317724
#>  [16,] -0.03518964
#>  [17,] -0.32983328
#>  [18,] -1.10343264
#>  [19,] -0.20185031
#>  [20,]  0.64635073
#>  [21,]  1.91019519
#>  [22,] -0.18980434
#>  [23,]  1.71146983
#>  [24,] -1.30739014
#>  [25,]  1.42441960
#>  [26,] -1.41499224
#>  [27,] -0.98966592
#>  [28,] -1.01325664
#>  [29,] -5.03333267
#>  [30,] -1.92139187
#>  [31,]  0.08071337
#>  [32,]  0.08843648
#>  [33,]  0.02241534
#>  [34,]  3.35556385
#>  [35,] -3.74777529
#>  [36,]  3.54964426
#>  [37,] -3.77616124
#>  [38,]  3.28110014
#>  [39,]  2.19794615
#>  [40,] -1.30868558
#>  [41,]  0.13023440
#>  [42,] -1.67587745
#>  [43,]  0.23801624
#>  [44,] -4.46047326
#>  [45,] -5.52007279
#>  [46,]  1.60618440
#>  [47,]  4.35296746
#>  [48,] -0.83035361
#>  [49,]  2.21355287
#>  [50,] -0.37041542
#>  [51,] -3.48999499
#>  [52,] -1.60297708
#>  [53,]  0.04088990
#>  [54,]  1.39462006
#>  [55,]  2.86584670
#>  [56,] -0.91096781
#>  [57,]  2.87994143
#>  [58,] -0.78050119
#>  [59,]  0.46602766
#>  [60,]  1.18075487
#>  [61,] -5.56030850
#>  [62,]  0.43052711
#>  [63,]  0.19651578
#>  [64,]  0.71505277
#>  [65,]  1.68448848
#>  [66,] -1.36147012
#>  [67,]  4.43791384
#>  [68,] -1.17121728
#>  [69,]  1.98856488
#>  [70,]  1.97044237
#>  [71,] -1.70829064
#>  [72,]  3.84086720
#>  [73,]  0.98852392
#>  [74,] -2.46214338
#>  [75,]  1.40242692
#>  [76,]  3.24178272
#>  [77,]  2.99017618
#>  [78,]  0.37210363
#>  [79,] -0.35102827
#>  [80,] -1.89997532
#>  [81,]  2.39701756
#>  [82,]  0.02552688
#>  [83,]  0.96600898
#>  [84,] -0.69275829
#>  [85,]  3.52273367
#>  [86,]  1.16113938
#>  [87,]  3.54188481
#>  [88,]  2.21497720
#>  [89,] -0.21116221
#>  [90,] -0.52084470
#>  [91,]  2.11046081
#>  [92,] -5.90574331
#>  [93,] -1.33042559
#>  [94,]  3.25251364
#>  [95,]  4.28984602
#>  [96,]  1.91667365
#>  [97,]  3.80574057
#>  [98,] -4.53201128
#>  [99,]  1.57251950
#> [100,]  2.78680556
#> [101,] -0.04228833
#> [102,]  0.30246072
#> [103,]  0.31433539
#> [104,] -2.70686726
#> [105,] -1.75453026
#> [106,] -0.48193849
#> [107,]  4.85630257
#> [108,] -1.87090590
#> 
#> [[2]]
#>                 [,1]
#>   [1,] -0.4235486997
#>   [2,]  1.5102855655
#>   [3,] -1.2984175425
#>   [4,]  0.4251644842
#>   [5,]  1.6678019318
#>   [6,]  0.9289811633
#>   [7,]  0.0604550323
#>   [8,] -1.0486605467
#>   [9,]  0.0406692510
#>  [10,] -0.0575040327
#>  [11,]  0.7757798982
#>  [12,]  0.5368666601
#>  [13,] -0.6103770022
#>  [14,] -1.1227176353
#>  [15,] -0.3922994254
#>  [16,] -2.4777200622
#>  [17,] -1.4949042527
#>  [18,] -0.2229388894
#>  [19,]  0.1593653617
#>  [20,] -0.0421748236
#>  [21,]  0.1397179039
#>  [22,]  1.2961242731
#>  [23,] -0.6651532780
#>  [24,] -0.1720859307
#>  [25,]  0.0850723177
#>  [26,] -0.0837358022
#>  [27,]  1.2615015016
#>  [28,] -1.7752395961
#>  [29,]  1.5539763197
#>  [30,] -0.0875773958
#>  [31,] -1.2806771298
#>  [32,] -1.5848352910
#>  [33,]  0.2352175473
#>  [34,] -0.0200579111
#>  [35,] -1.1350840261
#>  [36,] -0.4249342211
#>  [37,]  0.9757674220
#>  [38,] -1.3484493107
#>  [39,] -0.9859583630
#>  [40,] -0.2806132513
#>  [41,]  0.3818362211
#>  [42,] -0.4457643468
#>  [43,]  0.9841896304
#>  [44,]  1.2975027640
#>  [45,] -0.7520861929
#>  [46,] -0.1236866016
#>  [47,]  0.5787528687
#>  [48,] -0.9582753579
#>  [49,] -0.0003258788
#>  [50,] -1.2808778126
#>  [51,]  0.1894689491
#>  [52,]  1.0903449566
#>  [53,] -0.0849940300
#>  [54,]  0.1909236950
#>  [55,]  0.3072263161
#>  [56,] -1.2951649002
#>  [57,] -0.5811148503
#>  [58,] -1.8741465887
#>  [59,]  0.3262724807
#>  [60,] -0.6534414360
#>  [61,] -0.2171123170
#>  [62,]  0.2845864799
#>  [63,]  0.6948030720
#>  [64,]  0.3982704850
#>  [65,] -0.9639040072
#>  [66,] -1.2879642229
#>  [67,] -0.3695047282
#>  [68,]  1.0169080528
#>  [69,] -0.5868881115
#>  [70,] -0.6824698242
#>  [71,]  2.2542474979
#>  [72,]  1.4378613553
#>  [73,] -0.5493545999
#>  [74,] -0.1553413733
#>  [75,]  0.9365420661
#>  [76,]  0.3538831401
#>  [77,] -0.6312242170
#>  [78,]  0.5013370059
#>  [79,]  0.5758257668
#>  [80,] -0.7852467612
#>  [81,]  0.7672272666
#>  [82,]  0.7466489999
#>  [83,]  0.3235898201
#>  [84,] -0.0625010660
#>  [85,] -0.3540571125
#>  [86,] -0.7100556217
#>  [87,]  0.4212746334
#>  [88,]  1.0772388185
#>  [89,] -1.4874758525
#>  [90,]  0.3891847331
#>  [91,]  1.5973185832
#>  [92,]  0.6415259212
#>  [93,]  0.6254344757
#>  [94,] -0.0980782180
#>  [95,]  0.5141852776
#>  [96,]  0.4978313303
#>  [97,] -0.9622746379
#>  [98,]  0.6251096698
#>  [99,] -0.7733073461
#> [100,]  0.3245572755
#> 
#> [[3]]
#>             [,1]
#>  [1,] -1.1167511
#>  [2,] -0.1764798
#>  [3,]  0.4426426
#>  [4,]  1.0588583
#>  [5,] -1.5133822
#>  [6,]  1.3995358
#>  [7,] -0.9781066
#>  [8,]  0.6256796
#>  [9,] -0.3245021
#> [10,]  0.8798102
#> 
#> [[4]]
#>               [,1]
#>   [1,] -0.28169871
#>   [2,]  0.13834713
#>   [3,] -0.14232212
#>   [4,] -0.85460416
#>   [5,] -0.19449976
#>   [6,] -0.42642368
#>   [7,]  0.36901460
#>   [8,] -0.41528603
#>   [9,] -0.62092081
#>  [10,]  0.12564406
#>  [11,] -1.91865645
#>  [12,] -0.73294597
#>  [13,]  0.74094920
#>  [14,] -0.89879656
#>  [15,] -0.52146547
#>  [16,] -2.01505722
#>  [17,]  0.31403033
#>  [18,] -0.22711473
#>  [19,]  1.10984432
#>  [20,]  1.63643256
#>  [21,]  0.29851854
#>  [22,]  1.15679871
#>  [23,]  0.68812383
#>  [24,]  1.19712586
#>  [25,] -0.19955553
#>  [26,]  1.05674191
#>  [27,]  0.37405285
#>  [28,]  0.19455775
#>  [29,] -0.71141553
#>  [30,] -0.07029605
#>  [31,]  0.58218463
#>  [32,] -1.28493014
#>  [33,]  0.75649574
#>  [34,]  1.56364225
#>  [35,]  0.15233866
#>  [36,] -0.09606163
#>  [37,]  0.20034096
#>  [38,]  1.71143067
#>  [39,] -1.25410058
#>  [40,] -0.19375509
#>  [41,] -0.15736454
#>  [42,] -0.69533985
#>  [43,]  0.26137998
#>  [44,]  0.87005833
#>  [45,] -0.32365127
#>  [46,]  0.58317018
#>  [47,] -0.03985992
#>  [48,]  0.33535786
#>  [49,]  1.36308761
#>  [50,] -0.60527503
#>  [51,] -0.44590560
#>  [52,] -0.53820263
#>  [53,]  1.49014860
#>  [54,]  0.42064858
#>  [55,]  0.29130766
#>  [56,] -0.49619024
#>  [57,] -0.03137256
#>  [58,]  0.31365579
#>  [59,]  0.12628647
#>  [60,]  0.17752312
#>  [61,] -0.77694182
#>  [62,]  1.39015124
#>  [63,]  1.18024708
#>  [64,] -0.16437773
#>  [65,]  0.37876528
#>  [66,]  0.19872473
#>  [67,] -1.79469669
#>  [68,] -1.10778396
#>  [69,] -0.57548603
#>  [70,] -1.55236762
#>  [71,]  0.31637198
#>  [72,] -1.76184905
#>  [73,]  1.00737619
#>  [74,]  0.19967214
#>  [75,] -1.03781863
#>  [76,]  0.41662137
#>  [77,] -0.54747319
#>  [78,]  1.35927669
#>  [79,] -0.27766910
#>  [80,] -0.30588483
#>  [81,]  0.22704932
#>  [82,] -0.83577168
#>  [83,] -1.50098670
#>  [84,] -1.34514857
#>  [85,]  1.23525119
#>  [86,] -0.68929897
#>  [87,] -0.88396166
#>  [88,]  1.62855303
#>  [89,]  1.15903820
#>  [90,]  0.71145446
#>  [91,]  0.26830787
#>  [92,]  1.57252640
#>  [93,]  0.20881390
#>  [94,]  0.47727785
#>  [95,]  0.48786238
#>  [96,] -0.72180994
#>  [97,] -1.02156879
#>  [98,]  1.32034865
#>  [99,] -0.70450417
#> [100,] -0.02234409
#> 
#> [[5]]
#>               [,1]
#>  [1,]  0.007066982
#>  [2,]  0.023309834
#>  [3,]  0.418313522
#>  [4,]  0.152096353
#>  [5,]  0.558903587
#>  [6,]  0.098324603
#>  [7,] -0.068197773
#>  [8,] -0.149056524
#>  [9,] -0.162754089
#> [10,] -0.008606841
#> 
#> [[6]]
#>            [,1]
#> [1,]  0.2885944
#> [2,] -0.7325983
#>