Initialize input vectors for the ACMTF algorithm

Initialize input vectors for the ACMTF algorithm

Usage

initializeACMTF(Z, numComponents, initialization = "random", output = "Fac")

Arguments

Z

List object as generated by setupCMTFdata().

numComponents

Integer stating the number of desired components for the CMTF model.

initialization

Initialization method, either "random" or "nvec" (default "random"). Random will initialize random input vectors. Nvec will initialize vectors according to an singular value decomposition of the (matricized, if needed) concatenated datasets per mode.

output

How to return output: as a "Fac" object (default) or vectorized ("vect").

Value

List or vector of initialized input vectors per mode.

Examples

A = array(rnorm(108*2), c(108, 2))
B = array(rnorm(100*2), c(100, 2))
C = array(rnorm(10*2), c(10, 2))
D = array(rnorm(100*2), c(100,2))
E = array(rnorm(10*2), c(10,2))

df1 = reinflateTensor(A, B, C)
df2 = reinflateTensor(A, D, E)
datasets = list(df1, df2)
modes = list(c(1,2,3), c(1,4,5))
Z = setupCMTFdata(datasets, modes, normalize=FALSE)

init = initializeACMTF(Z, 2)