Cross-validation of ACMTF-R by jack-knifing

Cross-validation of ACMTF-R by jack-knifing

Usage

ncrossreg(
  Z,
  Y,
  maxNumComponents = 5,
  alpha = 1,
  beta = rep(0.001, length(Z$object)),
  epsilon = 1e-08,
  pi = 0.5,
  method = "CG",
  cg_update = "HS",
  line_search = "MT",
  max_iter = 10000,
  max_fn = 10000,
  abs_tol = 1e-10,
  rel_tol = 1e-10,
  grad_tol = 1e-10,
  nstart = 5,
  numCores = 1
)

Arguments

Z

Combined dataset and mode object as produced by setupCMTFdata().

Y

Dependent variable (regression part).

maxNumComponents

Maximum number of components to investigate (default 5).

alpha

Scalar penalizing the components to be norm 1 (default 1).

beta

Vector of penalty values for each dataset, penalizing the lambda terms (default 1e-3).

epsilon

Scalar value to make it possible to compute the partial derivatives of lambda (default 1e-8).

pi

Pi value of the loss function as specified by Van der Ploeg et al., 2025.

method

Optimization method to use (default = "CG", the conjugate gradient). See mize::mize() for other options.

cg_update

Update method for the conjugate gradient algorithm, see mize::mize() for the options (default="HS", Hestenes-Steifel).

line_search

Line search algorithm to use, see mize::mize() for the options (default="MT", More-Thuente).

max_iter

Maximum number of iterations.

max_fn

Maximum number of function evaluations.

abs_tol

Function tolerance criterion for convergence.

rel_tol

Relative function tolerance criterion for convergence.

grad_tol

Absolute tolerence for the l2-norm of the gradient vector.

nstart

Number of randomly initialized models to create per number of components (default 5).

numCores

Number of cores to use (default 1). Setting this to higher than 1 will make the algorithm run in parallel.

Value

A plot showcasing the variance explained in the X blocks and Y as well as the RMSE of Y.

Examples

set.seed(123)
A = array(rnorm(25*2), c(25, 2))
B = array(rnorm(100*2), c(100, 2))
C = array(rnorm(10*2), c(10, 2))
D = array(rnorm(100*2), c(100, 2))
E = array(rnorm(10*2), c(10, 2))

df1 = reinflateTensor(A, B, C)
df2 = reinflateTensor(A, D, E)
datasets = list(df1, df2)
modes = list(c(1,2,3), c(1,4,5))
Z = setupCMTFdata(datasets, modes)
Y = matrix(A[,1])

# Uses poor settings to make the example work.
plot = ncrossreg(Z, Y, maxNumComponents=2, max_iter=2, nstart=2)
#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if
#> `.name_repair` is omitted as of tibble 2.0.0.
#> ℹ Using compatibility `.name_repair`.
#> ℹ The deprecated feature was likely used in the CMTFtoolbox package.
#>   Please report the issue to the authors.